Move auth-options unverified email test to the mocked stack
