+++ example/src/androidTest/resources/fetch-auth-options-response-success-unverified-email.json
@@ -1,0 +1,1 @@
+{"passwordless":false,"email_verified":false}
